back-end/requirements.txt
asgiref
crispy-bootstrap5
distlib
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
importlib_metadata
Markdown
pillow
platformdirs
psycopg2
sqlparse
typing_extensions
tzdata
virtualenv
zipp
